ml/requirements.txt
fastapi
uvicorn
transformers
torch
clickhouse-driver
numpy
python-dotenv